Adverse Events Timeline Panel Server

Adverse Events Timeline Panel Server

Usage

mod_AETimeline_Server(
  id,
  rctv_dfAE_Study,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  chrDateFields = c(mincreated_dts = "AE Entry Date", aest_dt = "AE Start Date", aeen_dt
    = "AE End Date")
)

Arguments

id

character The id for this element.

rctv_dfAE_Study

reactive dataframe A shiny::reactive() object that returns a study-level adverse events domain dataframe.

rctv_strGroupID

reactive character A shiny::reactiveVal() object that returns the GroupID of the selected group (usually site), and can be used to update which group is selected.

rctv_strGroupLevel

reactive character A shiny::reactiveVal() object that returns the selected GroupLevel.

rctv_strSubjectID

reactive character A shiny::reactive() object that returns the SubjectID of the selected participant.

chrDateFields

character A named vector of fields in dfAE to treat as dates, where the names are the column names and the values are what to display to users. Unnamed elements will be used as both the field name and the display name. These fields are currently used in the Timeline panel.

Value

A module server function.