AE Plugin Definition

Plugins for gsm.app are defined by a named list with information about the plugin. These can be loaded from a yaml file with gsm.app::plugin_Read(). We provide a ready-to-use plugin definition for this AE plugin.

Usage

pluginAE()

Value

A plugin definition for use in gsm.app.

Examples

pluginAE()
#> $meta
#> $meta$Name
#> [1] "Adverse Events"
#> 
#> 
#> $shiny
#> $shiny$UI
#> function (id) 
#> {
#>     ns <- NS(id)
#>     bslib::navset_card_underline(id = ns("selected_tab"), title = "Adverse Events", 
#>         bslib::nav_panel("Dashboard", mod_AEDashboard_UI(ns("dashboard"))), 
#>         bslib::nav_panel("Explorer", aeExplorerOutput(ns("explorer"))), 
#>         bslib::nav_panel("Timeline", aeTimelinesOutput(ns("timeline"))), 
#>         full_screen = TRUE)
#> }
#> <bytecode: 0x562976b48400>
#> <environment: namespace:gsm.ae>
#> 
#> $shiny$Server
#> function (id, rctv_dfAE, rctv_dfSUBJ, rctv_strSiteID, rctv_strSubjectID) 
#> {
#>     moduleServer(id, function(input, output, session) {
#>         rctv_dfAE_mod <- reactive({
#>             dfAE <- rctv_dfAE()
#>             if (NROW(dfAE)) {
#>                 dfAE <- dfAE %>% dplyr::inner_join(rctv_dfSUBJ(), 
#>                   by = "SubjectID") %>% dplyr::mutate(aetoxgr = toupper(sub("^out of bound:", 
#>                   "", .data$aetoxgr)), aest_dy = .data$aest_dt - 
#>                   .data$firstparticipantdate + 1L, aeen_dy = .data$aeen_dt - 
#>                   .data$firstparticipantdate + 1L) %>% dplyr::arrange(.data$SubjectID, 
#>                   .data$aest_dy, .data$aeen_dy, .data$mdrpt_nsv, 
#>                   .data$aetoxgr) %>% dplyr::mutate(seq = dplyr::row_number(), 
#>                   .by = "SubjectID")
#>             }
#>         })
#>         aes_settings <- list(id_col = "SubjectID", seq_col = "seq", 
#>             stdy_col = "aest_dy", endy_col = "aeen_dy", term_col = "mdrpt_nsv", 
#>             bodsys_col = "mdrsoc_nsv", severity_col = "aetoxgr", 
#>             serious_col = "aeser")
#>         mod_AEDashboard_Server("dashboard", rctv_dfAE_mod)
#>         output$explorer <- render_aeExplorer({
#>             dfAE <- rctv_dfAE_mod()
#>             if (NROW(dfAE)) {
#>                 render_SC_Widget(lData = list(dm = rctv_dfSUBJ(), 
#>                   aes = dfAE), lSettings = list(dm = list(id_col = "SubjectID"), 
#>                   aes = aes_settings), fnInit = safetyCharts::init_aeExplorer, 
#>                   strWidgetName = "aeExplorer")
#>             }
#>         })
#>         output$timeline <- render_aeTimelines({
#>             dfAE <- rctv_dfAE_mod()
#>             if (NROW(dfAE)) {
#>                 render_SC_Widget(lData = dfAE, lSettings = aes_settings, 
#>                   fnInit = safetyCharts::init_aeTimelines, strWidgetName = "aeTimelines")
#>             }
#>         })
#>     })
#> }
#> <bytecode: 0x562976b4aac8>
#> <environment: namespace:gsm.ae>
#> 
#> 
#> $domains
#> [1] "AE"   "SUBJ"
#>