AE Plugin Definition

Plugins for gsm.app are defined by a named list with information about the plugin. These can be loaded from a yaml file with gsm.app::plugin_Read(). We provide a ready-to-use plugin definition for this AE plugin.

Usage

pluginAE(
  strField_Created = "mincreated_dts",
  strMetricID_AE = "Analysis_kri0001",
  strMetricID_SAE = "Analysis_kri0002"
)

Arguments

strField_Created

length-1 character The field that indicates when the adverse event was logged. Defaults to "mincreated_dts".

strMetricID_AE

length-1 character The ID of the adverse events metric. Defaults to "Analysis_kri0001".

strMetricID_SAE

length-1 character The ID of the serious adverse events metric. Defaults to "Analysis_kri0001".

Value

A plugin definition for use in gsm.app.

Examples

pluginAE()
#> $meta
#> $meta$Name
#> [1] "Adverse Events"
#> 
#> 
#> $shiny
#> $shiny$UI
#> function (id) 
#> {
#>     ns <- NS(id)
#>     bslib::page_fillable(mod_AEDashboard_UI(ns("dashboard")))
#> }
#> <bytecode: 0x555e5f2e3af0>
#> <environment: namespace:gsm.ae>
#> 
#> $shiny$Server
#> function (id, dfAnalyticsInput, dfResults, rctv_dSnapshotDate, 
#>     rctv_strGroupID, rctv_strGroupLevel, rctv_strSubjectID, strMetricID_AE = "Analysis_kri0001", 
#>     strMetricID_SAE = "Analysis_kri0002") 
#> {
#>     dfAnalyticsInput <- dfAnalyticsInput %>% dplyr::filter(.data$MetricID %in% 
#>         c(strMetricID_AE, strMetricID_SAE)) %>% dplyr::mutate(MetricID = dplyr::case_match(.data$MetricID, 
#>         strMetricID_AE ~ "AE", strMetricID_SAE ~ "SAE"), dplyr::across(c("Numerator", 
#>         "Denominator"), as.integer)) %>% dplyr::arrange(.data$GroupLevel, 
#>         .data$GroupID, .data$SubjectID, .data$MetricID, .data$SnapshotDate)
#>     moduleServer(id, function(input, output, session) {
#>         rctv_dSnapshotDatePrevious <- reactive({
#>             dSnapshotDates <- as.Date(unique(dfResults$SnapshotDate))
#>             dSnapshotDatesPrevious <- dSnapshotDates[!is.na(dSnapshotDates) & 
#>                 dSnapshotDates < rctv_dSnapshotDate()]
#>             if (length(dSnapshotDatesPrevious)) {
#>                 return(max(dSnapshotDatesPrevious, na.rm = TRUE))
#>             }
#>             return(as.Date(NA, origin = "1970-01-01"))
#>         })
#>         rctv_strGroupID_inferred <- reactive({
#>             strGroupID <- rctv_strGroupID()
#>             strSubjectID <- rctv_strSubjectID()
#>             if (is.null(strGroupID) && length(strSubjectID) == 
#>                 1) {
#>                 test1 <- dfAnalyticsInput$SubjectID == strSubjectID
#>                 test2 <- dfAnalyticsInput$GroupLevel == rctv_strGroupLevel()
#>                 return(dfAnalyticsInput$GroupID[test1 & test2][[1]])
#>             }
#>             return(strGroupID)
#>         }) %>% bindCache(rctv_strSubjectID(), rctv_strGroupID(), 
#>             rctv_strGroupLevel())
#>         mod_AEDashboard_Server("dashboard", dfAnalyticsInput = dfAnalyticsInput, 
#>             rctv_dSnapshotDate = rctv_dSnapshotDate, rctv_dSnapshotDatePrevious = rctv_dSnapshotDatePrevious, 
#>             rctv_strGroupID = rctv_strGroupID_inferred, rctv_strGroupLevel = rctv_strGroupLevel, 
#>             rctv_strSubjectID = rctv_strSubjectID)
#>     })
#> }
#> <bytecode: 0x555e5f2e4848>
#> <environment: namespace:gsm.ae>
#> 
#> 
#> $spec
#> $spec$AE
#> $spec$AE$aest_dt
#> $spec$AE$aest_dt$type
#> [1] "Date"
#> 
#> 
#> $spec$AE$aeser
#> $spec$AE$aeser$type
#> [1] "character"
#> 
#> 
#> $spec$AE$SubjectID
#> $spec$AE$SubjectID$type
#> [1] "character"
#> 
#> 
#> $spec$AE$mincreated_dts
#> $spec$AE$mincreated_dts$type
#> [1] "character"
#> 
#> 
#> 
#> $spec$SUBJ
#> $spec$SUBJ$SubjectID
#> $spec$SUBJ$SubjectID$type
#> [1] "character"
#> 
#> 
#> 
#> 
#> $config
#> $config$strField_Created
#> [1] "mincreated_dts"
#> 
#> $config$strMetricID_AE
#> [1] "Analysis_kri0001"
#> 
#> $config$strMetricID_SAE
#> [1] "Analysis_kri0002"
#> 
#> 
#> $packages
#> list()
#> 
#> $required_inputs
#> character(0)
#> 
#> $workflows
#> list()
#>