Parameters used in multiple functions

Reused parameter definitions are gathered here for easier usage. Note: We intentionally use id rather than the more standardized strID throughout this package, because shiny::testServer() specifically expects an id argument in module server functions.

Arguments

id

character The id for this element.

dfAnalyticsInput

data.frame Participant-level metric data.

dfResults

data.frame A stacked summary of analysis pipeline output.

rctv_dSnapshotDate

reactive Date A shiny::reactive() object that returns the date of a data snapshot.

rctv_dSnapshotDatePrevious

reactive Date A shiny::reactive() object that returns the date of a data snapshot immediately before rctv_dSnapshotDate, or NA if there aren't any snapshots before the selected date.

rctv_dfAE

reactive dataframe A shiny::reactive() object that returns an adverse events domain dataframe.

rctv_dfSUBJ

reactive dataframe A shiny::reactive() object that returns a subject domain dataframe.

rctv_strGroupID

reactive character A shiny::reactiveVal() object that returns the GroupID of the selected group (usually site), and can be used to update which group is selected.

rctv_strGroupLevel

reactive character A shiny::reactiveVal() object that returns the selected GroupLevel.

rctv_strSubjectID

reactive character A shiny::reactive() object that returns the SubjectID of the selected participant.

session

environment Session from which to make a child scope (the default should almost always be used).

strField_Created

length-1 character The field that indicates when the adverse event was logged. Defaults to "mincreated_dts".

strField_GroupActive

length-1 character The field that indicates when the group (usually site or country) became active. Defaults to "site_active_dt".

strField_StudyActive

length-1 character The field that indicates when the study became active. Defaults to "act_fpfv".

strMetricID_AE

length-1 character The ID of the adverse events metric. Defaults to "Analysis_kri0001".

strMetricID_SAE

length-1 character The ID of the serious adverse events metric. Defaults to "Analysis_kri0001".